### 2025-04-01 #7973 @editorialbot
[PRE REVIEW]: yieldplotlib: A unified library for exoplanet yield code visualizations
**Submitting author:** @steigersg (<a href="http://orcid.org/0000-0002-4787-3285">Sarah Steiger</a>)
**Repository:** https://github.com/HWO-Yield-Visualizations/yieldplotlib
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
