### 2025-04-01 #7973 @warrickball
Hi @steigersg, thanks for your submission! All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. I'll indicate up front that this will probably take several months given the current backlog. Sorry, and thanks for your patience!

In the meantime, note that [our docs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) ask that papers are short:

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

Though I won't insist on the paper being strictly under 1000 words, it would be good to try to get it below about 1200 or so. The main point of the paper is to describe the context and purpose of the software. Though this can be motivated with examples, the paper shouldn't double up on documentation. In this case, I think it's mainly the "Methods and Functionality" section that you should try to reduce.
