### 2025-04-01 #7973 @warrickball
Editors, note that about 5000 lines of Python is in `src/yieldplotlib/key_map.py`.
