### 2025-04-02 42a000e
docs: add functionality.md file moving implementation details from paper to docs.
