### 2025-04-02 6e42b4c
docs: first pass at shortening paper
