+++ paper/paper.md
@@ -52,9 +52,7 @@ comparing results difficult. The need for a unified library for visualizing the 
 and outputs of these yield codes in a complete, descriptive, and accessible way has therefore
 also become apparent. To this end we have developed `yieldplotlib`, an open-source Python
 library to communicate the results of yield codes to the broader community and produce
-publication-quality plots. Currently, there are modules for analyzing AYO and EXOSIMS,
-but `yieldplotlib` is easily extensible and support for other yield codes can be easily
-added in the future.
+publication-quality plots.
 
 # Statement of need
 
@@ -91,45 +89,20 @@ the future.
 ## Parsing and Getting Values
 `yieldplotlib` provides a loading system with a unified interface for accessing data from
 the yield codes. The system manages the complex and inconsistent file structures of the
-AYO and EXOSIMS inputs and outputs by organizing them into a hierarchical tree of nodes
-representing files and directories. This abstraction creates a consistent API that
+AYO and EXOSIMS inputs and outputs by creating a consistent API that
 allows users to query data without needing to understand and parse the underlying
 data products. For collaboration purposes, the valid queries are managed in a Google
 Sheet where collaborators have linked the EXOSIMS and AYO keys to a universal key in `yieldplotlib`.
-This Google Sheet is processed into a `key_map` which is updated daily. Users can update
-the sheet, download it as a CSV file, and process it locally for development.
+This Google Sheet is processed into a `key_map` which is updated daily.
 An example excerpt from the CSV file can be found in
 \autoref{fig:key_map_csv}.
 
 ![Example portion of the `yieldplotlib` key map CSV file containing the
 mappings between AYO, EXOSIMS, and `yieldplotlib` parameters.\label{fig:key_map_csv}](figures/ypl_csv_table.jpeg)
 
-Once the yield packages are loaded and parsed, a getter can be called on the directory objects
-to return the corresponding value from the respective yield code, for example:
-
-```angular2html
-from yieldplotlib.load import AYODirectory, EXOSIMSDirectory
-
-ayo = AYODirectory(Path("path/to/my/ayo_data"))
-exosims = EXOSIMSDirectory(Path("path/to/my/exosims_data"))
-
-ayo.get("yield_earth")
-exosims.get("yield_earth")
-```
-
 Yield input packages (YIPs) specifying input coronagraph parameters can also be loaded and accessed
 using the same file node and directory structure. This allows users to access key coronagraph performance
-metrics that serve as critical inputs to these yield codes. In order to process the YIPs, `yieldplotlib`
-uses `yippy`[^1] as a backend, though the user interface is identical to generating the AYO and EXOSIMS
-directories, for example:
-
-[^1]: github.com/CoreySpohn/yippy
-
-```angular2html
-from yieldplotlib.load import YIPDirectory
-
-yip = YIPDirectory(Path("path/to/my/yip_data"))
-```
+metrics that serve as critical inputs to these yield codes.
 
 ## Plotting
 
@@ -139,10 +112,7 @@ yip = YIPDirectory(Path("path/to/my/yip_data"))
 of the wide variety of customization options that  `matplotlib` offers, as well as the extensive
 knowledge base many users of `yieldplotlib` will have with that package. The `yieldplotlib` generic
 plots are used for single yield run visualizations and can make scatter plots, standard plots,
-and histograms. This extension is achieved through a function that runs when `yieldplotlib`
-is imported that automatically adds new plotting methods (prefixed with `ypl_`) to `matplotlib`'s
-`Axes` class, allowing users to directly call methods like `ax.ypl_plot()` and
-`ax.ypl_scatter()` on any `matplotlib` axes object.
+and histograms.
 
[…]
