+++ docs/user/functionality.md
@@ -6,8 +6,7 @@ the yield codes. The system manages the complex and inconsistent file structures
 AYO and EXOSIMS inputs and outputs by organizing them into a hierarchical tree of nodes
 representing files and directories. This abstraction creates a consistent API that
 allows users to query data without needing to understand and parse the underlying
-data products. For collaboration purposes, the valid queries are managed in a Google
-Sheet where collaborators have linked the EXOSIMS and AYO keys to a universal key in `yieldplotlib`.
+data products. For collaboration purposes, the valid queries are managed in a [Google Sheet](https://docs.google.com/spreadsheets/d/1qZ-JH0xnYSqeWOrirEIvJ-NcoMVu90pDgeRsqbDUcxM/edit?usp=sharing) where collaborators have linked the EXOSIMS and AYO keys to a universal key in `yieldplotlib`.
 This Google Sheet is processed into a `key_map` which is updated daily. Users can update
 the sheet, download it as a CSV file, and process it locally for development.
 
