### 2025-04-02 d31b9d0
docs: Shorten paper more
