+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'yieldplotlib: A unified library for exoplanet yield code visualizations'
+title: "yieldplotlib: A unified library for exoplanet yield code visualizations"
 tags:
   - Python
   - astronomy
@@ -23,14 +23,14 @@ authors:
     affiliation: "1, 3, 4"
 
 affiliations:
- - name: Goddard Space Flight Center, United States
-   index: 1
- - name: Space Telescope Science Institute, United States
-   index: 2
- - name: The Catholic University of America, United States
-   index: 3
- - name: Center for Research and Exploration in Space Science and Technology, NASA/GSFC, United States
-   index: 4
+  - name: Goddard Space Flight Center, United States
+    index: 1
+  - name: Space Telescope Science Institute, United States
+    index: 2
+  - name: The Catholic University of America, United States
+    index: 3
+  - name: Center for Research and Exploration in Space Science and Technology, NASA/GSFC, United States
+    index: 4
 
 date: 1 April 2025
 bibliography: paper.bib
@@ -62,36 +62,35 @@ Worlds Observatory (HWO; @HWOFeinberg2024). As HWO is being developed and trade 
 explored, yield codes such as the Altruistic Yield Optimizer (AYO; @AYO2014) and EXOSIMS
 [@EXOSIMS2016] that can calculate the expected number of detected and characterized planets for a
 given mission architecture are essential. While these yield codes have the same goal, they are
-complex, written in different programming languages (AYO is written in IDL and
-EXOSIMS is written in Python), calculate yield with different methods, and have
+complex, written in different programming languages (IDL and
+Python), calculate yield with different methods, and have
 major differences in their inputs and outputs.
 The need for a unified library for visualizing these yield codes in a complete, descriptive,
-and accessible way has therefore also become apparent. This is non-trivial due to the differing
+and accessible way has become apparent from cross model validation efforts. This is non-trivial due to the differing
 methods, syntaxes, structures, and assumptions that each of these codes make.
 
-Despite the challenges, when these values are interrogated directly, new insights are achieved.
-Some of these insights are highlighted in detail in @ETCCrossCal2025 where a comparison of just
-the internal exposure time calculations of AYO and EXOSIMS revealed sources of previously unknown
-discrepancy. `yieldplotlib` is a continuation of that initial work, but aimed instead at the higher
-level yield products (such as the total number of detected exo-Earths) which are of the most direct
-interest.
+Despite the challenges, interrogating the values has already provided valuable insights.
+The study @ETCCrossCal2025 compared just
+the internal exposure time calculations of AYO and EXOSIMS and revealed previously unknown
+discrepancies. `yieldplotlib` is a more ambitious continuation of that work, capable of accessing more
+important yield products such as the exo-Earth yield.
 
 To visualize the inputs and outputs of AYO and EXOSIMS, `yieldplotlib` uses a custom loading and
-parsing structure that allows for the easy access of equivalent data across each of the codes.
+parsing structure to easily access equivalent data across both codes.
 This allows `yieldplotlib` to communicate the results of yield codes to the broader community
-and produce publication-quality plots without the need to understand the complex underlying yield
-codes that were used to generate the data. Currently `yieldplotlib` contains modules for analyzing
-AYO and EXOSIMS, but is easily extensible and support for other yield codes can be easily added in
-the future.
+and produce publication-quality plots without manually processing the complex underlying data.
+Currently `yieldplotlib` contains modules for analyzing
+AYO and EXOSIMS, but can be easily extended to support other yield codes.
 
 # Methods and Functionality
 
 ## Parsing and Getting Values
+
[…]
