### 2025-04-02 4958863
docs: Another pass on the paper
