+++ paper/paper.md
@@ -38,42 +38,27 @@ bibliography: paper.bib
 
 # Summary
 
-NASA’s next flagship observatory, as recommended by the Astro2020 decadal survey,
-is the Habitable World’s Observatory (HWO) which has the ambitious goal to "search
-for biosignatures from a robust number of about ~25 habitable zone planets and be
-a transformative facility for general astrophysics". The total number of habitable zone planets
-detected is referred to as the exo-Earth "yield" and estimating it will be critically important
-to HWO's success. As HWO is being developed
-and trade spaces are explored, yield codes such as the Altruistic Yield Optimizer (AYO)
-and EXOSIMS that can calculate the expected number of detected and characterized planets
-for a given mission architecture are essential. While these yield codes have the same goal,
-they can be complex and have major differences in their inputs and outputs which makes
-comparing results difficult. The need for a unified library for visualizing the inputs
-and outputs of these yield codes in a complete, descriptive, and accessible way has therefore
-also become apparent. To this end we have developed `yieldplotlib`, an open-source Python
-library to communicate the results of yield codes to the broader community and produce
-publication-quality plots.
+NASA’s next flagship observatory, the Habitable Worlds Observatory (HWO), aims
+to detect and charaterize ~25 habitable zone planets. The total number of
+habitable zone planets detected is referred to as the exo-Earth "yield" and
+accurate yield estimates will be critically important to the mission's success. Tools
+like the Altruistic Yield Optimizer (AYO) and EXOSIMS provide these yield
+estimates but differ in language, methods, and outputs. `yieldplotlib` provides
+a unified library that can visualize the inputs and outputs of these yield
+codes in a complete, descriptive, and accessible way.
 
 # Statement of need
 
-Expected exoplanet yield is an important metric when evaluating the success of proposed
-flagship space observatory architectures such as those currently being considered for the Habitable
-Worlds Observatory (HWO; @HWOFeinberg2024). As HWO is being developed and trade spaces are
-explored, yield codes such as the Altruistic Yield Optimizer (AYO; @AYO2014) and EXOSIMS
-[@EXOSIMS2016] that can calculate the expected number of detected and characterized planets for a
-given mission architecture are essential. While these yield codes have the same goal, they are
-complex, written in different programming languages (IDL and
-Python), calculate yield with different methods, and have
-major differences in their inputs and outputs.
-The need for a unified library for visualizing these yield codes in a complete, descriptive,
-and accessible way has become apparent from cross model validation efforts. This is non-trivial due to the differing
-methods, syntaxes, structures, and assumptions that each of these codes make.
+To evaluate different designs for HWO [@HWOFeinberg2024], yield codes such as AYO [@AYO2014] and EXOSIMS
+[@EXOSIMS2016] calculate the expected exo-Earth yield for each architecture.
+While these yield codes have the same goal, their implementations
+are so different that validation has become a major obstacle for the community.
 
-Despite the challenges, interrogating the values has already provided valuable insights.
-The study @ETCCrossCal2025 compared just
+Previous efforts have already provided valuable insights. The study @ETCCrossCal2025 compared
 the internal exposure time calculations of AYO and EXOSIMS and revealed previously unknown
-discrepancies. `yieldplotlib` is a more ambitious continuation of that work, capable of accessing more
-important yield products such as the exo-Earth yield.
+discrepancies. `yieldplotlib` is a more ambitious continuation of that work.
+It is a Python library capable of easily accessing hundreds of important quantities
[…]
