### 2025-04-02 d4c9324
docs: small paper updates
