### 2025-04-02 0f9c479
docs: add JOSS submitted badge
