### 2025-04-07 90e6c55
fix: Add tabulate install to workflow
