+++ .github/workflows/generate_parameters_docs.yml
@@ -34,7 +34,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install pandas google-auth google-auth-oauthlib google-api-python-client
+        pip install pandas google-auth google-auth-oauthlib google-api-python-client tabulate
 
     - name: Generate parameter table from Google Sheets
       env:
