+++ .pre-commit-config.yaml
@@ -6,7 +6,7 @@ repos:
       - id: name-tests-test
       - id: end-of-file-fixer
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.12.0
+    rev: v0.12.2
     hooks:
       # Run the linter.
       - id: ruff
