### 2025-08-06 0862a77
chore: Updating workflow to v6 of setup-uv
