+++ .github/workflows/tests.yml
@@ -24,6 +24,7 @@ jobs:
       with:
         python-version: ${{ matrix.python-version }}
         enable-cache: true
+        activate-environment: true
 
     - name: Install nox
       run: uv tool install nox
