+++ .github/workflows/draft-pdf.yaml
@@ -15,7 +15,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout repository
-        uses: actions/checkout@v4
+        uses: actions/checkout@v5
         with:
           token: ${{ secrets.HWO_RELEASE_PLEASE_TOKEN }}
       - name: Build draft PDF
