+++ .github/workflows/generate_parameters_docs.yml
@@ -27,7 +27,7 @@ jobs:
         token: ${{ secrets.HWO_RELEASE_PLEASE_TOKEN }}
 
     - name: Set up Python
-      uses: actions/setup-python@v5
+      uses: actions/setup-python@v6
       with:
         python-version: "3.x"
 
