### 2025-10-26 #7973 @warrickball
Hi @jrenaud90, @mauricemolli, @lorenaacuna & @nichollsh, would any of you be willing to review this submission to JOSS? As a reminder, we carry out our checklist-driven reviews here in GitHub issues and follow [these guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html).
