### 2025-11-25 #9401 @warrickball
For the record, @gbergsten has indicated that as of quite recently, he and @steigersg are both employed at STScI. I am happy to waive this because it's a large institution and there has been no past collaboration.

@gbergsten, don't forget to create a checklist here by commenting with `@editorialbot generate my checklist`, even if you don't expect to immediately progress through much of it.
