+++ .github/workflows/download-gsheet.yml
@@ -22,7 +22,7 @@ jobs:
 
     steps:
     - name: Checkout repository
-      uses: actions/checkout@v5
+      uses: actions/checkout@v6
       with:
         token: ${{ secrets.HWO_RELEASE_PLEASE_TOKEN }}
 
