### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#128 @ben-cassese
Git clone command issues
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3535455792

In [this section of the developer documentation](https://yieldplotlib.readthedocs.io/en/latest/user/dev.html#clone-the-repository), you list `git clone git@github.com:HWO-Yield-Visualizations/yieldplotlib.git` as the command for cloning the repository. This only works for users who have set up their SSH keys on Github, however, so I recommend using a more generic command: `git clone https://github.com/HWO-Yield-Visualizations/yieldplotlib.git`.
