### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#129 @ben-cassese
Reliance on removed yippy functionality
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3581487343

Towards the end of the Getting Started tutorial, you demonstrate how to use `plot_core_throughtput` from `yieldplotlib.plots.yip_plots`. However, this relies on `yip.coronagraph.get_throughput_curve` from yippy, which was removed in [this commit](https://github.com/CoreySpohn/yippy/commit/eec0d8cf49b8300febb77ee2b837168d2640a0cd) in April. I'd either pin an older version of yippy as a dependency or update `plot_core_throughtput` to use yippy's new interface.
