### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#130 @ben-cassese
Animation inline rendering error
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3535455792

In the Getting Started tutorial, you use `ani = make_offax_psf_movie(yip, "offax_psf_animation.gif")`. This successfully saves a .gif animation, but leaves `ani` as `None` since `make_offax_psf_movie` has no return. This causes the next line, `HTML(ani.to_jshtml())`,  to error and for the animation not to play within a jupyter notebook. I'm guessing you just need to add `return ani` after [this line](https://github.com/HWO-Yield-Visualizations/yieldplotlib/blob/1d198294172e6691e0a162e5d903482c705eae6c/src/yieldplotlib/plots/yip_plots.py#L67), but double check that. Note that the online version of the docs also currently shows the same error, so after a fix, be sure to regenerate those.
