### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#131 @ben-cassese
Accessibility testing results
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3535455792

At the very end of the Getting Started tutorial, you demonstrate how to use yieldplotlib's `AccessibilityManager`. Right now the `run_checks()` method sends its output to the logger as a WARNING. Since earlier in the tutorial you set the logging level to ERROR, nothing gets printed in a jupyter notebook after running `run_checks()`. Users can still see the results by running `am.warnings`, but I'd either a) add an explicit mention not to expect any printed output to the tutorial or b) add a print statement to `run_checks` so that users see something no matter how they've configured the logger.
