### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#132 @ben-cassese
JOSS requirement for Community Guidelines
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3535455792

Though there is a great docs page for developers wishing to contribute to the code, there is technically no documentation that satisfies the JOSS requirements "2. Report issues or problems with the software 3. Seek support". A quick sentence on the landing page of the docs would be enough for these, just let users know if they should use the Github page or email the maintainers if they have issues.
