### 2025-12-02 HWO-Yield-Visualizations/yieldplotlib#133 @ben-cassese
Google sheet access
Comment associated with JOSS review https://github.com/openjournals/joss-reviews/issues/9401#issuecomment-3535455792

The Google Sheets page that hosts the live list of valid queries is not publicly viewable (I get sent to the standard "request access" Google page). This is not impeding anyone who wants to use what's already there since the key_map.py file is still present in the repository, but since you state that users can update the sheet and/or download it for local manipulation, either the permissions on the sheet should be changed or the docs should be updated with instructions on how to request access.
