### 2025-12-03 70d03ae
docs: add Google sheet access instructions.
