### 2025-12-03 3df0c06
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
