### 2025-12-03 7278b46
docs: add instructions if users encounter issues with the code
