### 2025-12-03 4927944
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
