### 2025-12-03 382c21d
docs: update git cloning instructions to not use SSH keys
