### 2025-12-04 #9401 @gbergsten
Excellent stuff, all! The manuscript, docs and code are extremely easy to follow, and the need/motivation/intended use are clear. This is an exciting software poised to save the community a ton of time thanks to the cross-simulator compatibility (more on that below), and help explore the pesky disagreements between said simulators. I ran into the same issues as @ben-cassese while running through the tutorial, most of which (pending [issue #129 on removed yippy functionality](https://github.com/HWO-Yield-Visualizations/yieldplotlib/issues/129)) have already been resolved. I'm still poking at other parts of the software and will report back with any issues if I find them!
