### 2025-12-04 a276ccf
fix: update throughput calculation in plot_core_throughtput function
