### 2025-12-04 2d359d5
feat: Add searchable common parameter table to the tutorial
