### 2025-12-09 #9401 @steigersg
Hi @warrickball, the latest release (v1.1.0) is the most up to date. The Zenodi doi is 10.5281/zenodo.15013340 for **all versions** and 10.5281/zenodo.17822644 for **v1.1.0 specifically** (not sure which one you need).

I also confirmed that all ORCIDs and affiliations are correct and that the Zenodo license matches the GitHub license.

Thanks so much for your assistance and let us know if you need anything else at this time!
