### 2025-12-09 #9401 @warrickball
At an initial glance, the paper looks good enough that it should only require some minor tweaks that I'll make via PR, hopefully in the next few days. I'll run the draft publication step now just to see if that presents anything I'll need to fix as AEiC too.
