### 2025-12-09 #9401 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
