### 2025-12-11 #9401 @warrickball
Thanks once again to @ben-cassese and @gbergsten for reviewing this submission! JOSS simply wouldn't be possible without its community of volunteers and I personally really appreciate your exemplary engagement with the process.

Congratulations @steigersg and @CoreySpohn, your paper has been published in JOSS! Thanks too for staying so engaged, and I apologise again that there was such a long wait before we could get started.
